+++ NetIdMobileSdk/src/main/java/de/netid/mobile/sdk/appauth/AppAuthManager.kt
@@ -63,6 +63,13 @@ internal interface AppAuthManager {
     fun getAccessToken(): String?
 
     /**
+     * Sets an access token, overwriting the actual token in the SDK.
+     * This makes most sense after initializing the SDK and getting a token.
+     * @param accessToken access token to set, of type ``String``
+     */
+    fun setAccessToken(accessToken: String?)
+
+    /**
      * Returns the currently available permission token if there is one, null otherwise.
      * @return permissionToken
      */