+++ NetIdMobileSdk/src/main/java/de/netid/mobile/sdk/appauth/AppAuthManagerImpl.kt
@@ -31,12 +31,13 @@ import org.json.JSONException
 import org.json.JSONObject
 import java.util.concurrent.locks.ReentrantLock
 
+
 internal class AppAuthManagerImpl(context: Context) : AppAuthManager {
 
     companion object {
         private const val scheme = "https://"
         private const val scopePermissionManagement = "permission_management"
-        private val browserDenyList =  BrowserDenyList(
+        private val browserDenyList = BrowserDenyList(
             VersionedBrowserMatcher.FIREFOX_BROWSER,
             VersionedBrowserMatcher.FIREFOX_CUSTOM_TAB
         )
@@ -50,11 +51,7 @@ internal class AppAuthManagerImpl(context: Context) : AppAuthManager {
     private var authorizationServiceConfiguration: AuthorizationServiceConfiguration? = null
     private var authService: AuthorizationService? = null
     private var authState: AuthState? = null
-    private var sharedPreferences: SharedPreferences
-
-    init {
-        sharedPreferences = context.getSharedPreferences(STORE_NAME, Context.MODE_PRIVATE)
-    }
+    private var sharedPreferences: SharedPreferences = context.getSharedPreferences(STORE_NAME, Context.MODE_PRIVATE)
 
     override fun getAccessToken(): String? {
         return getAuthState()?.accessToken
@@ -67,11 +64,21 @@ internal class AppAuthManagerImpl(context: Context) : AppAuthManager {
     }
 
     override fun getAuthState(): AuthState? {
-        if (authState != null)
+        if (authState != null) {
             return authState
+        }
 
         authState = readState()
         return authState
+    }
+
+    override fun setAccessToken(accessToken: String?) {
+        val lastTokenResponse = getAuthState()?.lastTokenResponse
+        if (lastTokenResponse != null) {
+            val builder = TokenResponse.Builder(lastTokenResponse.request).setAccessToken(accessToken)
+            val modifiedTokenResponse = builder.build()
+            getAuthState()?.update(modifiedTokenResponse, null)
+        }
     }
 
     /**